updated reducers and seperated orders actions
