+++ src/actions/auth.js
@@ -33,22 +33,6 @@ export const authError = error => ({
     error
 });
 
-export const ORDER_FAIL = 'ORDER_FAIL';
-export const orderFail = error => ({
-    type: AUTH_ERROR,
-    error
-});
-
-export const ORDER_SUCCESS = 'ORDER_SUCCESS'
-export const orderSuccess = order => ({
-    type: ORDER_SUCCESS,
-    order
-});
-export const DELETE_SUCCESS = 'DELETE_SUCCESS';
-export const deleteOrder = deleteSuccess => ({
-    type: DELETE_SUCCESS,
-    deleteSuccess
-});
 
 // Stores the auth token in state and localStorage, and decodes and stores
 // the user data stored in the token
@@ -116,43 +100,4 @@ export const refreshAuthToken = () => (dispatch, getState) => {
             dispatch(clearAuth());
             clearAuthToken(authToken);
         });
-};
-
-/*
-*/
-export const submitOrder = (order) => dispatch => {
-    return (
-        fetch(`${API_BASE_URL}/order`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                Authorization: `Bearer ${localStorage.getItem('authToken')}`
-            },
-            body: JSON.stringify(
-                order
-            )
-        })
-            // Reject any requests which don't return a 200 status, creating
-            // errors which follow a consistent format
-            .then(res => normalizeResponseErrors(res))
-            .then(res => dispatch(orderSuccess(res)))
-                //res.json(console.log("success lol make a dispatch")))
-            //
-            //action to display success console.log("success lol make a dispatch")
-            //.then(({authToken}) => storeAuthInfo(authToken, dispatch))
-            .catch(err => {
-                const {code} = err;
-                const message =
-                    code === 401
-                        ? 'Incorrect username or password'
-                        : 'Unable to submit order, please log in again';
-                dispatch(authError(err));
-                // Could not authenticate, so return a SubmissionError for Redux Form
-                return Promise.reject(
-                    new SubmissionError({
-                        _error: message
-                    })
-                );
-            })
-    );
 };
