+++ src/components/about.js
@@ -8,7 +8,8 @@ export default function About(){
 			<h1 className="aboutTitle">About</h1>
 			<p className="about">Superheroes Delivers! is a non sequitur way to deliver laughter to children
 			 and adults who are in the hospital for a prolong period of time. Chocolates and flowers are nice, 
-			 but they are better when delivered by a superhero. Superheroes Delivers! was founded by Rob Birch.</p>
+			 but they are better when delivered by a superhero. Superheroes Delivers! was founded by Rob Birch. 
+			 If you have issues please call (512) 515-1628.</p>
 		</div>
 		)
 }
