+++ src/components/app.css
@@ -6,7 +6,7 @@ body {
   background-color: white;
 }
 
-a{
+a {
 	text-decoration: none;
 	border-radius: 4px;
 	border: solid black 3px;
@@ -22,12 +22,12 @@ h2, h3{
 
 }
 
-h2{
+h2 {
 	font-size: 1.1em;
 	width: 20%;
 }
 
-h3{
+h3 {
 	font-size: 1.0em;
 	width: 25%;
 }
