+++ src/components/superhero.js
@@ -7,8 +7,9 @@ import Superman from '../images/Superman.jpg';
 // Create an img element which displays test.png
 //const image = <img src={TestImage} />;
 //if statement logic for batman, superman and spiderman
-
+//import store
 export default function Superhero(props){
+	console.log(props)
 	let heroPic;
 	if (props.hero === 'Batman') {
 		heroPic = Batman
