+++ src/config.js
@@ -1,5 +1,5 @@
 export const API_BASE_URL =
-    process.env.REACT_APP_API_BASE_URL || 'https://superherodeliverserver.herokuapp.com/api';
+    process.env.REACT_APP_API_BASE_URL || 'http://localhost:8080/api';
     //below are two different urls for local host api
     //1 http://localhost:8080/api 
     //2 'https://superherodeliverserver.herokuapp.com/api';
