+++ src/reducers/auth.js
@@ -3,25 +3,15 @@ import {
     CLEAR_AUTH,
     AUTH_REQUEST,
     AUTH_SUCCESS,
-    AUTH_ERROR,
-    ORDER_FAIL,
-    ORDER_SUCCESS,
-    /*REQUEST_ORDERS,
-    RECEIVE_ORDERS*/
+    AUTH_ERROR
 } from '../actions/auth';
 
-import {
-    REQUEST_ORDERS,
-    RECEIVE_ORDERS
-} from '../actions/orders'
 
 const initialState = {
     authToken: null, // authToken !== null does not mean it has been validated
     currentUser: null,
     loading: false,
-    error: null,
-    order: null,
-    orders: null
+    error: null
 };
 
 export default function reducer(state = initialState, action) {
@@ -49,25 +39,6 @@ export default function reducer(state = initialState, action) {
             loading: false,
             error: action.error
         });
-    } else if (action.type === ORDER_FAIL){
-        return Object.assign({}, state, {
-            error: action.error
-        });
-    } else if (action.type === ORDER_SUCCESS){
-        return Object.assign({}, state, {
-            order: action.order
-        });
-    } else if (action.type === REQUEST_ORDERS){ 
-        console.log(action,action.type)
-        return Object.assign({},state, {
-            order: action.orders,
-            loading: false
-        });
-    } else if (action.type === RECEIVE_ORDERS){
-        return Object.assign({},state,{
-            order: action.order,
-            loading: true
-        })
-    }
+    } 
     return state;
 }
